In progress design refactor
I pulled out all the graph parts of circuit and made CircuitEdge, CircuitGraph
and CircuitNode. Rationale is that for fast deletes, you need both in edges and
out edges anyway, so to keep them in sync, we delegate that to another class so
we shouldn't have to worry about it apart from the graph classes.

An alternative is a lazy delete, but idk